requirements.txt
streamlit==1.32.0
openai==0.28.1
python-dotenv==1.0.0
requests==2.28.2
python-dateutil==2.8.2
bcrypt==4.0.1 